+++ next.config.ts
@@ -8,10 +8,6 @@ const nextConfig: NextConfig = {
         protocol: 'https',
         hostname: 'scontent.cdninstagram.com',
       },
-      {
-        protocol: 'https',
-        hostname: 'fakestoreapi.com',
-      },
     ],
   },
 };
